+++ satochip-lib/src/main/java/org/satochip/client/ApplicationStatus.java
@@ -96,7 +96,10 @@ public class ApplicationStatus {
     public byte getPin0RemainingCounter(){
         return PIN0_remaining_tries;
     }
-  
+    public byte getPuk0RemainingCounter(){
+        return PUK0_remaining_tries;
+    }
+
     public String toString(){
         String status_info=   "setup_done: " + setup_done + "\n"+
                                   "is_seeded: " + is_seeded + "\n"+
@@ -113,6 +116,15 @@ public class ApplicationStatus {
                 ((int) protocol_minor_version * (1 << 16)) +
                 ((int) applet_major_version * (1 << 8)) +
                 ((int) applet_minor_version);
+    }
+
+    public String getCardVersionString() {
+        String version_string =
+            protocol_major_version + "." + 
+            protocol_minor_version + "-" +
+            applet_major_version + "." + 
+            applet_minor_version;
+        return version_string;
     }
 
     public int getProtocolVersion() {