+++ satochip-lib/src/main/java/org/satochip/client/SatochipCommandSet.java
@@ -453,7 +453,8 @@ public class SatochipCommandSet {
         data[offset++] = (byte) 0x01;
 
         APDUCommand plainApdu = new APDUCommand(cla, ins, p1, p2, data);
-        logger.info("SATOCHIPLIB: C-APDU cardSetup:" + plainApdu.toHexString());
+        //logger.info("SATOCHIPLIB: C-APDU cardSetup:" + plainApdu.toHexString());
+        logger.info("SATOCHIPLIB: C-APDU cardSetup");
         APDUResponse respApdu = this.cardTransmit(plainApdu);
         logger.info("SATOCHIPLIB: R-APDU cardSetup:" + respApdu.toHexString());
 
@@ -468,6 +469,24 @@ public class SatochipCommandSet {
         return respApdu;
     }
 
+    public APDUResponse cardSendResetCommand() throws Exception {
+        byte[] data = new byte[]{};
+
+        APDUCommand plainApdu = new APDUCommand(
+                0xB0,
+                INS_RESET_TO_FACTORY,
+                0x00,
+                0x00,
+                data
+        );
+
+        // reset command must be sent in clear, without other commands interferring between reset commands
+        logger.warning("SATOCHIPLIB: C-APDU cardSendResetCommand:" + plainApdu.toHexString());
+        APDUResponse respApdu = apduChannel.send(plainApdu);
+        logger.warning("SATOCHIPLIB: R-APDU cardSendResetCommand:" + respApdu.toHexString());
+
+        return respApdu;
+    }
 
     /****************************************
      *             PIN MGMT                  *
@@ -477,19 +496,125 @@ public class SatochipCommandSet {
         System.arraycopy(pin, 0, this.pin0, 0, pin.length);
     }
 
-    public APDUResponse cardVerifyPIN() {
+    public APDUResponse cardVerifyPIN(byte[] pin) throws Exception {
 
-        if (pin0 == null) {
-            // TODO: specific exception
-            throw new RuntimeException("PIN required!");
+        byte[] mypin = pin;
+        if (mypin == null){
+            if (pin0 == null) {
+                // TODO: specific exception
+                throw new RuntimeException("PIN required!");
+            }
+            mypin = pin0;
         }
 
-        APDUCommand plainApdu = new APDUCommand(0xB0, INS_VERIFY_PIN, 0x00, 0x00, pin0);
-        logger.info("SATOCHIPLIB: C-APDU cardVerifyPIN:" + plainApdu.toHexString());
-        APDUResponse respApdu = this.cardTransmit(plainApdu);
-        logger.info("SATOCHIPLIB: R-APDU cardVerifyPIN:" + respApdu.toHexString());
+        try {
+            APDUCommand plainApdu = new APDUCommand(0xB0, INS_VERIFY_PIN, 0x00, 0x00, mypin);
+            //logger.info("SATOCHIPLIB: C-APDU cardVerifyPIN:" + plainApdu.toHexString());
+            logger.info("SATOCHIPLIB: C-APDU cardVerifyPIN");
+            APDUResponse rapdu = this.cardTransmit(plainApdu);
+            logger.info("SATOCHIPLIB: R-APDU cardVerifyPIN:" + rapdu.toHexString());
 
-        return respApdu;
+            rapdu.checkAuthOK();
+            this.pin0 = mypin; // cache new pin
+            return rapdu;
+
+        } catch (WrongPINException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (WrongPINLegacyException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (BlockedPINException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (APDUException e){
+            this.pin0 = null;
+            throw e;
+        } catch (Exception e){
+            this.pin0 = null;
+            throw e;
+        }
+    }
+
+    public APDUResponse cardVerifyPIN() throws Exception {
+        return cardVerifyPIN(this.pin0);
+    }
+
+    public APDUResponse cardChangePin(byte[] oldPin, byte[] newPin) throws Exception {
+        logger.info("SATOCHIPLIB: changeCardPin START");
+
+        byte[] data = new byte[1 + oldPin.length + 1 + newPin.length];
+        data[0] = (byte) oldPin.length;
+        System.arraycopy(oldPin, 0, data, 1, oldPin.length);
+        data[1 + oldPin.length] = (byte) newPin.length;
+        System.arraycopy(newPin, 0, data, 2 + oldPin.length, newPin.length);
+        setPin0(newPin);
+        try{
+            APDUCommand plainApdu = new APDUCommand(0xB0, INS_CHANGE_PIN, 0x00, 0x00, data);
+            //logger.info("SATOCHIPLIB: C-APDU changeCardPin:"+ plainApdu.toHexString());
+            logger.info("SATOCHIPLIB: C-APDU changeCardPin");
+            APDUResponse rapdu = this.cardTransmit(plainApdu);
+            logger.info("SATOCHIPLIB: R-APDU changeCardPin:"+ rapdu.toHexString());
+            
+            rapdu.checkAuthOK();
+            return rapdu;
+
+        } catch (WrongPINException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (WrongPINLegacyException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (BlockedPINException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (APDUException e){
+            this.pin0 = null;
+            throw e;
+        } catch (Exception e){
+            this.pin0 = null;
+            throw e;
+        }
+    }
+
+    public APDUResponse cardUnblockPin(byte[] puk) throws Exception {
+        APDUCommand plainApdu = new APDUCommand(
+                0xB0,
+                INS_UNBLOCK_PIN,
+                0x00,
+                0x00,
+                puk
+        );
+
+        try{
+            //logger.info("SATOCHIPLIB: C-APDU cardUnblockPin:" + plainApdu.toHexString());
+            logger.info("SATOCHIPLIB: C-APDU cardUnblockPin");
+            APDUResponse rapdu = this.cardTransmit(plainApdu);
+            logger.info("SATOCHIPLIB: R-APDU cardUnblockPin:" + rapdu.toHexString());
+
+            rapdu.checkAuthOK();
+            return rapdu;
+
+        } catch (WrongPINException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (WrongPINLegacyException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (BlockedPINException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (ResetToFactoryException e) {
+            this.pin0 = null;
+            throw e;
+        } catch (APDUException e){
+            this.pin0 = null;
+            throw e;
+        } catch (Exception e){
+            this.pin0 = null;
+            throw e;
+        }
+        
     }
 
     /****************************************
@@ -500,7 +625,8 @@ public class SatochipCommandSet {
         //TODO: check seed (length...)
         APDUCommand plainApdu = new APDUCommand(0xB0, INS_BIP32_IMPORT_SEED, masterseed.length, 0x00, masterseed);
 
-        logger.info("SATOCHIPLIB: C-APDU cardBip32ImportSeed:" + plainApdu.toHexString());
+        //logger.info("SATOCHIPLIB: C-APDU cardBip32ImportSeed:" + plainApdu.toHexString());
+        logger.info("SATOCHIPLIB: C-APDU cardBip32ImportSeed");
         APDUResponse respApdu = this.cardTransmit(plainApdu);
         logger.info("SATOCHIPLIB: R-APDU cardBip32ImportSeed:" + respApdu.toHexString());
 
@@ -1232,7 +1358,8 @@ public class SatochipCommandSet {
                     chunk
             );
 
-            logger.warning("SATOCHIPLIB: C-APDU seedkeeperImportSecret:" + plainApdu.toHexString());
+            //logger.warning("SATOCHIPLIB: C-APDU seedkeeperImportSecret:" + plainApdu.toHexString());
+            logger.warning("SATOCHIPLIB: C-APDU seedkeeperImportSecret");
             respApdu = this.cardTransmit(plainApdu);
             logger.warning("SATOCHIPLIB: R-APDU seedkeeperImportSecret:" + respApdu.toHexString());
             respApdu.checkOK();
@@ -1261,7 +1388,8 @@ public class SatochipCommandSet {
                 chunk
         );
 
-        logger.warning("SATOCHIPLIB: C-APDU seedkeeperImportSecret:" + plainApdu.toHexString());
+        //logger.warning("SATOCHIPLIB: C-APDU seedkeeperImportSecret:" + plainApdu.toHexString());
+        logger.warning("SATOCHIPLIB: C-APDU seedkeeperImportSecret");
         respApdu = this.cardTransmit(plainApdu);
         logger.warning("SATOCHIPLIB: R-APDU seedkeeperImportSecret:" + respApdu.toHexString());
         respApdu.checkOK();
@@ -1346,7 +1474,6 @@ public class SatochipCommandSet {
         List<Byte> secretBytes = new ArrayList<>();
         int sigSize = 0;
         List<Byte> sigBytes = new ArrayList<>();
-//        byte[] sigBytes;
         while (true) {
             plainApdu = new APDUCommand(
                     0xB0,
@@ -1355,9 +1482,10 @@ public class SatochipCommandSet {
                     (byte) 0x02,
                     new byte[0]
             );
+            //logger.info("SATOCHIPLIB: C-APDU seedkeeperExportSecret:" + plainApdu.toHexString());
             logger.info("SATOCHIPLIB: C-APDU seedkeeperExportSecret:" + plainApdu.toHexString());
             respApdu = this.cardTransmit(plainApdu);
-            logger.info("SATOCHIPLIB: R-APDU seedkeeperExportSecret:" + respApdu.toHexString());
+            logger.info("SATOCHIPLIB: R-APDU seedkeeperExportSecret");
             respApdu.checkOK();
 
 
@@ -1655,29 +1783,6 @@ public class SatochipCommandSet {
         return sw == 0x9000;
     }
 
-    public Boolean changeCardPin(byte[] oldPin, byte[] newPin) throws Exception {
-        logger.info("SATOCHIPLIB: changeCardPin START");
-
-        byte[] data = new byte[1 + oldPin.length + 1 + newPin.length];
-        data[0] = (byte) oldPin.length;
-        System.arraycopy(oldPin, 0, data, 1, oldPin.length);
-        data[1 + oldPin.length] = (byte) newPin.length;
-        System.arraycopy(newPin, 0, data, 2 + oldPin.length, newPin.length);
-        setPin0(newPin);
-        try{
-            APDUCommand plainApdu = new APDUCommand(0xB0, INS_CHANGE_PIN, 0x00, 0x00, data);
-            logger.info("SATOCHIPLIB: C-APDU changeCardPin:"+ plainApdu.toHexString());
-            APDUResponse respApdu = this.cardTransmit(plainApdu);
-            logger.info("SATOCHIPLIB: R-APDU changeCardPin:"+ respApdu.toHexString());
-            int sw = respApdu.getSw();
-            return sw == 0x9000;
-        } catch (Exception e){
-            setPin0(oldPin);
-            return false;
-        }
-    }
-
-
     /****************************************
     *            PKI commands              *
     ****************************************/  
@@ -1808,7 +1913,6 @@ public class SatochipCommandSet {
         try{
             // load certs
             InputStream isCa = this.getClass().getClassLoader().getResourceAsStream("cert/ca.cert");  
-            //TODO: load subca cert depending on card type
             InputStream isSubca;
             if (cardType.equals("satochip")) {
                 isSubca = this.getClass().getClassLoader().getResourceAsStream("cert/subca-satochip.cert");